requirements.txt
torch>=1.11.0
torchvision>=0.12.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.64.0
scipy>=1.9.0
scikit-learn>=1.1.0
pyyaml
torchmetrics
tensorboard
pycocotools 